requirements.txt
# for python3.6
flask==2.0.2
flask-cors
pillow==8.4.0
tensorflow==2.5.0
# Jinja2==2.11.3
requests==2.26.0 # Used in testing.py


# for python 3.11
flask==2.3.1
flask-cors
pillow==9.3.0
tensorflow==2.14.0
# Jinja2==2.11.3
requests==2.26.0 # Used in testing.py